Add(opts): split up GuildOptions
The class was quite unreadable. Split it up into readable chunks
that are composed of the longest option stuff (birthday/autorole(s)/usermessage)
This should make (internal) discoverability and mantainance easier
aswell, since it was quite painful to read GuildOptions
