+++ src/main/java/net/kodehawa/mantarobot/options/GeneralOptions.java
@@ -94,6 +94,53 @@ public class GeneralOptions extends OptionHandler {
             ctx.sendLocalized("options.modlog_whitelist.success", EmoteReference.CORRECT, unBlacklisted);
         });
 
+        registerOption("modlog:blacklistwords:add", "Modlog Word Blacklist add", """
+                Adds a word to the modlog word blacklist (won't add any messages with that word). Can contain spaces.
+                **Example:** `~>opts modlog blacklistwords add mood`
+                """, "Sets the join message.", (ctx, args) -> {
+            if (args.length == 0) {
+                ctx.sendLocalized("options.modlog_blacklistwords_add.no_word", EmoteReference.ERROR);
+                return;
+            }
+
+            DBGuild dbGuild = ctx.getDBGuild();
+            GuildData guildData = dbGuild.getData();
+
+            if (guildData.getModLogBlacklistWords().size() > 20) {
+                ctx.sendLocalized("options.modlog_blacklistwords_add.too_many", EmoteReference.ERROR);
+                return;
+            }
+
+            String word = String.join(" ", args);
+            guildData.getModLogBlacklistWords().add(word);
+            dbGuild.save();
+            ctx.sendLocalized("options.modlog_blacklistwords_add.success", EmoteReference.CORRECT, word);
+        });
+
+        registerOption("modlog:blacklistwords:remove", "Modlog word blacklist remove", """
+                Removes a word from the modlog word blacklist. Can contain spaces
+                **Example:** `~>opts modlog blacklistwords remove mood`
+                """, "Sets the join message.", (ctx, args) -> {
+            if (args.length == 0) {
+                ctx.sendLocalized("options.modlog_blacklistwords_add.no_word", EmoteReference.ERROR);
+                return;
+            }
+
+            DBGuild dbGuild = ctx.getDBGuild();
+            GuildData guildData = dbGuild.getData();
+
+            String word = String.join(" ", args);
+
+            if (!guildData.getModLogBlacklistWords().contains(word)) {
+                ctx.sendLocalized("options.modlog_blacklistwords_remove.not_in", EmoteReference.ERROR, word);
+                return;
+            }
+
+            guildData.getModLogBlacklistWords().remove(word);
+            dbGuild.save();
+            ctx.sendLocalized("options.modlog_blacklistwords_remove.success", EmoteReference.CORRECT, word);
+        });
+
         registerOption("imageboard:tags:blacklist:add", "Blacklist imageboard tags", 
                 "Blacklists the specified imageboard tag from being looked up.",
                 "Blacklist imageboard tags", (ctx, args) -> {
